encapsulate into a component, basically working all the way now
