+++ packages/grafana-ui/src/components/MatchersUI/utils.ts
@@ -192,51 +192,3 @@ export function useSelectOptions(
     return options;
   }, [displayNames, currentName, firstItem, fieldType, baseNameMode]);
 }
-
-/**
- * @internal
- */
-export function useScopesOptions(
-  uniqScopes: Set<MatcherScope>,
-  currentScope?: MatcherScope
-): Array<ComboboxOption<MatcherScope>> {
-  return useMemo(() => {
-    // Remove the series scope from the set, so we can gaurantee it's the first option, and also
-    // because it's the default scope, so if it's the only one detected, we should not show the scope selector.
-    uniqScopes.delete('series');
-
-    const scopeNotFound = currentScope && currentScope !== 'series' && !uniqScopes.has(currentScope);
-
-    if (uniqScopes.size === 0 && !scopeNotFound) {
-      return [];
-    }
-
-    const arr: Array<ComboboxOption<MatcherScope>> = [
-      {
-        label: getGroupLabelForScope('series'),
-        description: getGroupDescriptionForScope('series'),
-        value: 'series',
-      },
-    ];
-
-    for (const scope of uniqScopes) {
-      arr.push({
-        label: getGroupLabelForScope(scope),
-        description: getGroupDescriptionForScope(scope),
-        value: scope,
-      });
-    }
-
-    if (scopeNotFound) {
-      const innerLabel = getGroupLabelForScope(currentScope);
-
-      arr.push({
-        label: t('grafana-ui.matchers.labels.not-found', '{{name}} (not found)', { name: innerLabel }),
-        description: getGroupDescriptionForScope(currentScope),
-        value: currentScope,
-      });
-    }
-
-    return arr;
-  }, [uniqScopes, currentScope]);
-}
